requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
flask>=3.0.0
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0
